% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcrt_causal_cov.R
\name{dCRT_causal_cov}
\alias{dCRT_causal_cov}
\title{CRAmed: A conditional randomization test for sparse and high-dimentional mediation analysis in microbiome data}
\usage{
dCRT_causal_cov(
  M_mat,
  Y,
  Exposure,
  cov,
  FDR = 0.05,
  n.times = 100,
  prefilter = TRUE,
  n.perm = 100,
  CI = FALSE,
  model = "gaussian",
  method = c("BH")
)
}
\arguments{
\item{M_mat}{a data frame or matrix containing the count data. Rows of the matrix represent observations and columns are the taxa.}

\item{Y}{a continuous outcome.}

\item{Exposure}{a binary exposure.}

\item{cov}{the matrix containing covariates.}

\item{FDR}{the threshold value for p-values.}

\item{n.times}{the number of times for sampling mediators from the ZINB model.}

\item{prefilter}{If TRUE, use lasso method to select variables.}

\item{n.perm}{the number of times for sampling dataset for calculating confidence interval.}

\item{CI}{If TRUE, calculate the confidence interval for NDE, NIE, NIEA, and NIEP.}

\item{model}{a model for fitting Y, including gaussian or binomial.}

\item{method}{the correction method for multiple testing.}
}
\value{
Return the identified mediators, corresponding p-values and confidence interval estimates of NDE, NIE, NIEA, and NIEP.
}
\description{
CRAmed: A conditional randomization test for sparse and high-dimentional mediation analysis in microbiome data
}
\examples{
otu_n <- 50;num <- 50
set.seed(123)
sim_zinb.mat <- sim_zinb(otu_n,num)
cramed.res <- dCRT_causal_cov(M_mat=sim_zinb.mat$M_mat,Y=sim_zinb.mat$Y_mat, Exposure=sim_zinb.mat$t,cov=matrix(rnorm(num)),n.perm=10, CI=TRUE)
}
