% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_zinb}
\alias{sim_zinb}
\title{Simulate the ZINB data}
\usage{
sim_zinb(otu_n, num)
}
\arguments{
\item{otu_n}{the number od taxa.}

\item{num}{the number of samples.}
}
\value{
Returns the data list, including the treatment, mediators and outcome.
}
\description{
Simulate the ZINB data
}
\examples{
otu_n <- 50;num <- 50
set.seed(123)
sim_zinb.mat <- sim_zinb(otu_n,num)
}
