% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_zinb}
\alias{sim_zinb}
\title{Simulate the ZINB data}
\usage{
sim_zinb(otu_n, num, alpha, beta, gamma, conf.num = NULL)
}
\arguments{
\item{otu_n}{the number od taxa.}

\item{num}{the number of samples.}

\item{alpha}{the corresponding regression coefficients for alpha.}

\item{beta}{the corresponding regression coefficients for beta.}

\item{gamma}{the corresponding regression coefficients for gamma.}

\item{conf.num}{the number of confounders.}
}
\value{
Returns the data list, including the treatment, mediators and outcome.
}
\description{
Simulate the ZINB data
}
\examples{
otu_n <- 100;num <- 50
set.seed(123)
sim_zinb.mat <- sim_zinb(otu_n, num, alpha=-2, beta=2, gamma=-2)
}
